XPO intersection_of processing
The python script to prepare the XPO needs to be integrated into the XPO build pipeline. We would benefit from a more detailed understanding of ROBOT and how it works with the transformation from OWL to OBO and specifically the java code that handles the intersection_of annotations.For now we are making do with the id_map.tsv as the data source.